add lson encoder to analysis dump
