+++ requirements.txt
@@ -1,4 +1,4 @@
-lightwood >= 1.1.0, < 1.2.0
+lightwood >= 1.2.0, < 1.3.0
 flask-restx >= 0.2.0
 flask >= 1.0, < 2.0
 python-multipart >= 0.0.5
